refactor: restore chart vue
